pyproject.toml
[tool.poetry]
name = "eaas-e2e-ma-mock"
version = "1.0.0"
description = "CloudBlue Connect Multi Account Extension for E2Es"
authors = ["CloudBlue LLC"]
license = "Apache-2.0"
packages = [
    { include = "e2e" },
]

[tool.poetry.dependencies]
python = ">=3.8,<4"
fastapi = "^0.78.0"
connect-eaas-core = "26.*"

[tool.poetry.plugins."connect.eaas.ext"]
"extension" = "e2e.events:E2EEventsApplication"
"webapp" = "e2e.webapp:E2EWebApplication"
